+++ types/gen/py/superblocks_types/utils/v1/utils_pb2.py
@@ -24,7 +24,7 @@ _sym_db = _symbol_database.Default()
 
 
 
-DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x14utils/v1/utils.proto\x12\x08utils.v1\"6\n\tSignature\x12\x15\n\x06key_id\x18\x01 \x01(\tR\x05keyId\x12\x12\n\x04\x64\x61ta\x18\x02 \x01(\x0cR\x04\x64\x61taB8Z6github.com/superblocksteam/agent/types/gen/go/utils/v1b\x06proto3')
+DESCRIPTOR = _descriptor_pool.Default().AddSerializedFile(b'\n\x14utils/v1/utils.proto\x12\x08utils.v1\"\xd1\x01\n\tSignature\x12\x15\n\x06key_id\x18\x01 \x01(\tR\x05keyId\x12\x12\n\x04\x64\x61ta\x18\x02 \x01(\x0cR\x04\x64\x61ta\x12\x1d\n\npublic_key\x18\x03 \x01(\x0cR\tpublicKey\x12;\n\talgorithm\x18\x04 \x01(\x0e\x32\x1d.utils.v1.Signature.AlgorithmR\talgorithm\"=\n\tAlgorithm\x12\x19\n\x15\x41LGORITHM_UNSPECIFIED\x10\x00\x12\x15\n\x11\x41LGORITHM_ED25519\x10\x01\x42\x38Z6github.com/superblocksteam/agent/types/gen/go/utils/v1b\x06proto3')
 
 _globals = globals()
 _builder.BuildMessageAndEnumDescriptors(DESCRIPTOR, _globals)
@@ -32,6 +32,8 @@ _builder.BuildTopDescriptorsAndMessages(DESCRIPTOR, 'utils.v1.utils_pb2', _globa
 if not _descriptor._USE_C_DESCRIPTORS:
   _globals['DESCRIPTOR']._loaded_options = None
   _globals['DESCRIPTOR']._serialized_options = b'Z6github.com/superblocksteam/agent/types/gen/go/utils/v1'
-  _globals['_SIGNATURE']._serialized_start=34
-  _globals['_SIGNATURE']._serialized_end=88
+  _globals['_SIGNATURE']._serialized_start=35
+  _globals['_SIGNATURE']._serialized_end=244
+  _globals['_SIGNATURE_ALGORITHM']._serialized_start=183
+  _globals['_SIGNATURE_ALGORITHM']._serialized_end=244
 # @@protoc_insertion_point(module_scope)