+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Superblocks partner ID to the Databricks integration
 - Fix scheduled job execution for signed jobs (when signature verification is enabled)
 - Now return Kinesis Plugin Metadata
+- Add signing key's corresponding public key and algorithm to the `Signature` proto message
 
 ## v1.15.1